services/ml/production/meteo.py
"""
Module pour l'analyse de l'impact météo sur la production.
"""

from typing import Dict, Any, List
from datetime import date
import numpy as np
from sqlalchemy.orm import Session
from sqlalchemy import and_

from .base import BaseProductionML
from models.production import Recolte
from services.weather_service import WeatherService

class MeteoAnalyzer(BaseProductionML):
    """Service ML pour l'analyse de l'impact météo"""
    
    def __init__(self, db: Session):
        super().__init__(db)
        self.weather_service = WeatherService(db)

    async def analyze_meteo_impact(
        self,
        parcelle_id: str,
        date_debut: date,
        date_fin: date
    ) -> Dict[str, Any]:
        """Analyse l'impact des conditions météo sur la production"""
        # Récupération des données
        meteo = await self.weather_service.get_historical_data(
            parcelle_id,
            date_debut,
            date_fin
        )
        
        recoltes = self.db.query(Recolte).filter(
            and_(
                Recolte.parcelle_id == parcelle_id,
                Recolte.date_recolte.between(date_debut, date_fin)
            )
        ).all()
        
        # Analyse des corrélations
        correlations = self._analyze_meteo_correlations(meteo, recoltes)
        
        # Identification des conditions critiques
        conditions_critiques = self._identify_critical_conditions(
            meteo,
            recoltes
        )
        
        return {
            "impact_score": self._calculate_impact_score(correlations),
            "correlations": correlations,
            "conditions_critiques": conditions_critiques,
            "recommandations": await self._generate_meteo_recommendations(
                parcelle_id,
                conditions_critiques
            )
        }

    def _analyze_meteo_correlations(
        self,
        meteo: List[Dict[str, Any]],
        recoltes: List[Recolte]
    ) -> Dict[str, float]:
        """Analyse les corrélations entre météo et rendements"""
        # TODO: Implémenter l'analyse des corrélations
        return {
            "temperature": 0.7,
            "humidite": 0.5,
            "precipitation": 0.3
        }

    def _identify_critical_conditions(
        self,
        meteo: List[Dict[str, Any]],
        recoltes: List[Recolte]
    ) -> List[Dict[str, Any]]:
        """Identifie les conditions météo critiques"""
        # TODO: Implémenter l'identification des conditions critiques
        return [
            {
                "condition": "Température élevée",
                "seuil": 30,
                "impact": -0.2,
                "frequence": 0.1
            },
            {
                "condition": "Précipitations faibles",
                "seuil": 10,
                "impact": -0.15,
                "frequence": 0.2
            }
        ]

    def _calculate_impact_score(
        self,
        correlations: Dict[str, float]
    ) -> float:
        """Calcule le score d'impact global"""
        return sum(correlations.values()) / len(correlations)

    async def _generate_meteo_recommendations(
        self,
        parcelle_id: str,
        conditions_critiques: List[Dict[str, Any]]
    ) -> List[str]:
        """Génère des recommandations basées sur l'analyse météo"""
        # TODO: Implémenter la génération de recommandations
        return [
            "Augmenter l'irrigation pendant les périodes de température élevée",
            "Planifier les récoltes en dehors des périodes de précipitations faibles"
        ]
